Make the code still buildable on older Mir
Uses MIR_SERVER_VERSION to conditionally include the new API in the mock
or stub only on the newer version. This makes the same codebase compiles
on both Mir 1.2 (our xenial), Mir 1.8 (our xenial_-_edge), and Mir 1.7
(in Ubuntu Focal, after additional commits).

Note that this uses a header for libmirserver-dev. However, rest
assured; I've checked Mir 2.2.0 in Hirsuite, <mir/version.h> is the only
header left in that package.