server: don't enforce libhybris wayland platform
Libhybris's EGL does not need to be in "wayland" platform to support
EGL extensions related to Wayland server. In fact, it can't, because
libhybris's "wayland" platform requires a wayland server, but we're
implementing a wayland server.

This fixes lomiri-system-compositor's spinner on Focal. The spinner is
forked directly from LSC and inherits its environment variables,
including EGL_PLATFORM. On Focal, we use libglvnd, which also reads
EGL_PLATFORM. This causes libglvnd to request a wrong platform window
from libhybris, breaking spinner.