include: Adapt AndroidDriverInterpreter interface
This interface is intended to be used by both the server platform
and the client platform to each integrate into the Android-side
GL/EGL driver stack respectively.

Some of these pure-virtual methods are not going to receive an
implementation anytime soon, they are there to be filled later
when someone finds the time to bring additional performance
improvements to mir-android-platform.

Other than that, adaptation of the implementations is following.