client: Make it possible to disallow EglFlush for certain programs
If the integrator chooses to do so they are now able to disallow
MirAndroidPlatformClientEglFlush & MirAndroidPlatformClientFenceSync
explicitly, by putting the list in quotes and prepending the entry
at hand with an exclamation mark. For example:

  MirAndroidPlatformClientEglFlush:
    - "!/usr/bin/unity8,!/usr/bin/lomiri,all"

This disallows EglFlush for Unity8/Lomiri, but keeps it for all
other programs. Order matters, so make sure to append the "all"
to the end of the list.