server: Allow flipping width and height for faulty hwcomposer HALs
The JingPad is a weird one. The Android hwcomposer HAL has width & height
in portrait mode, whereas the JingOS hwcomposer HAL has them in landscape.
This meant shipping the Android version as part of the device overlay.

Since I would prefer not having to overlay this HAL it needs a quirk property.
Introduce such a quirk property as a typical libdeviceinfo property setting.

Property: "MirAndroidPlatformServerFlipWidthAndHeight"
Type: boolean

This allows rendering contents on the screen using the default JingOS HAL.